v3

+++ srci/ar2rc.cpp
@@ -2,8 +2,8 @@
 #include "ar2rc.c"
 
 //Declarations
-const valarray<uint8_t> oktypes = {1,2,101,102};
-const size_t I = 1, O = 1;
+const valarray<size_t> oktypes = {1u,2u,101u,102u};
+const size_t I = 1u, O = 1u;
 size_t dim;
 
 //Description
@@ -29,7 +29,7 @@ struct arg_file  *a_fo = arg_filen("o","ofile","<file>",0,O,"output file (Y)");
 //Get options
 
 //Get dim
-if (a_d->count==0) { dim = (i1.R==1u) ? 1 : 0; }
+if (a_d->count==0) { dim = (i1.R==1u) ? 1u : 0u; }
 else if (a_d->ival[0]<0) { cerr << progstr+": " << __LINE__ << errstr << "dim must be nonnegative" << endl; return 1; }
 else { dim = size_t(a_d->ival[0]); }
 if (dim!=0 && dim!=1) { cerr << progstr+": " << __LINE__ << errstr << "dim must be 0 or 1" << endl; return 1; }
@@ -45,7 +45,7 @@ o1.R = i1.R; o1.C = i1.C; o1.S = i1.S; o1.H = i1.H;
 //Other prep
 
 //Process
-if (i1.T==1)
+if (i1.T==1u)
 {
     float *X, *Y;
     try { X = new float[i1.N()]; }
@@ -63,7 +63,7 @@ if (i1.T==1)
     }
     delete[] X; delete[] Y;
 }
-else if (i1.T==101)
+else if (i1.T==101u)
 {
     float *X, *Y;
     try { X = new float[2u*i1.N()]; }
