+++ srci/sig2ac.cpp
@@ -2,8 +2,8 @@
 #include "sig2ac.c"
 
 //Declarations
-const valarray<uint8_t> oktypes = {1,2,101,102};
-const size_t I = 1, O = 1;
+const valarray<size_t> oktypes = {1u,2u,101u,102u};
+const size_t I = 1u, O = 1u;
 size_t dim, L;
 char u;
 
@@ -45,7 +45,7 @@ else if (a_l->ival[0]<1) { cerr << progstr+": " << __LINE__ << errstr << "L (nla
 else { L = a_l->ival[0]; }
 
 //Get dim
-if (a_d->count==0) { dim = (i1.R==1u) ? 1 : 0; }
+if (a_d->count==0) { dim = (i1.R==1u) ? 1u : 0u; }
 else if (a_d->ival[0]<0) { cerr << progstr+": " << __LINE__ << errstr << "dim must be nonnegative" << endl; return 1; }
 else { dim = size_t(a_d->ival[0]); }
 if (dim!=0 && dim!=1) { cerr << progstr+": " << __LINE__ << errstr << "dim must be 0 or 1" << endl; return 1; }
@@ -56,19 +56,19 @@ u = (a_u->count>0);
 //Checks
 if (!i1.ismat()) { cerr << progstr+": " << __LINE__ << errstr << "input must be 1D or 2D" << endl; return 1; }
 if (i1.isempty()) { cerr << progstr+": " << __LINE__ << errstr << "input (X) found to be empty" << endl; return 1; }
-if (dim==0 && i1.R<2u) { cerr << progstr+": " << __LINE__ << errstr << "cannot work along a singleton dimension" << endl; return 1; }
-if (dim==1 && i1.C<2u) { cerr << progstr+": " << __LINE__ << errstr << "cannot work along a singleton dimension" << endl; return 1; }
+if (dim==0u && i1.R<2u) { cerr << progstr+": " << __LINE__ << errstr << "cannot work along a singleton dimension" << endl; return 1; }
+if (dim==1u && i1.C<2u) { cerr << progstr+": " << __LINE__ << errstr << "cannot work along a singleton dimension" << endl; return 1; }
 
 //Set output header info
 o1.F = i1.F; o1.T = i1.T;
-o1.R = (dim==0) ? uint32_t(L) : i1.R;
-o1.C = (dim==1) ? uint32_t(L) : i1.C;
+o1.R = (dim==0u) ? uint32_t(L) : i1.R;
+o1.C = (dim==1u) ? uint32_t(L) : i1.C;
 o1.S = i1.S; o1.H = i1.H;
 
 //Other prep
 
 //Process
-if (i1.T==1)
+if (i1.T==1u)
 {
     float *X, *Y;
     try { X = new float[i1.N()]; }
@@ -85,7 +85,7 @@ if (i1.T==1)
     }
     delete[] X; delete[] Y;
 }
-else if (i1.T==101)
+else if (i1.T==101u)
 {
     float *X, *Y;
     try { X = new float[2u*i1.N()]; }
